feat: login/register error handling
