+++ src/app/[locale]/components/Footer/Footer.tsx
@@ -1,8 +1,6 @@
 import Image from "next/image";
 import { useTranslations } from "next-intl";
 
-import "./footer.css";
-
 import logo from "../../../../../public/assets/images/logo.png";
 import {
   PiArrowUpBold,
