--- src/app/[locale]/components/Footer/footer.css
@@ -1,44 +1,0 @@
-.footerNewsletter {
-  display: flex;
-  align-items: center;
-}
-.footerNewsletter input {
-  padding: 0.7em;
-  border-radius: 20px;
-  border: solid 0.2px var(--strong-green);
-  width: 250px;
-  font-family: var(--montserrat);
-}
-
-.footerNewsletter button {
-  margin: 0.8em;
-  background-color: var(--strong-green);
-  padding: 0.8em;
-  width: 80px;
-  border: none;
-  border-radius: 25px;
-  color: var(--white);
-}
-
-.footerNewsletter button:hover {
-  text-decoration: underline;
-  cursor: pointer;
-}
-
-.upbutton a {
-  padding: 1em;
-  background-color: var(--strong-green);
-  color: white;
-  border-radius: 200px;
-}
-
-@media screen and (max-width: 768px) {
-  .footer {
-    gap: 20px;
-    justify-content: left;
-  }
-
-  .footerNewsletter {
-    display: block;
-  }
-}
