+++ src/app/[locale]/register/page.tsx
@@ -1,23 +1,21 @@
 import { useTranslations } from "next-intl";
 
-import Navbar from "../components/Navbar/Navbar";
 import Footer from "../components/Footer/Footer";
-import LoginForm from "../components/Login/LoginForm";
+import Navbar from "../components/Navbar/Navbar";
+import RegisterForm from "../components/Register/RegisterForm";
 
-import "../../../styles/login.css";
-
-const Login = () => {
-  const t = useTranslations("Login");
+const Register = () => {
+  const loginIdioms = useTranslations("Login");
   return (
     <>
       <Navbar />
       <section className="grid lg:grid-cols-2 items-center min-h-screen px-12 py-32 gap-12 bg-[url('/assets/images/imgLogin.png')] bg-cover bg-no-repeat bg-fixed">
         <h1 className="text-white text-4xl font-medium max-w-lg mx-auto lg:mx-0 text-balance text-center lg:text-start">
-          {t("login-title")}
+          {loginIdioms("login-title")}
         </h1>
 
-        <div className="w-full bg-white flex flex-col p-12 max-w-xl mx-auto rounded">
-          <LoginForm />
+        <div className="w-full bg-white flex flex-col py-4 px-6 lg:px-12 max-w-xl mx-auto rounded">
+          <RegisterForm />
         </div>
       </section>
       <Footer />
@@ -25,4 +23,4 @@ const Login = () => {
   );
 };
 
-export default Login;
+export default Register;
